superpower_testing/requirements.txt
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
pytest>=7.0.0
